Move API authorisation into controller superclass
